requirements.txt
pyinstaller
pandas
tqdm
faker
mysql-connector-python
python-dotenv
faker-commerce
